Fixing the issues related to changing the database

+++ RR-Common/src/main/java/com/scor/rr/exceptions/inuring/DeleteOfLastLayerException.java
@@ -4,7 +4,7 @@ import com.scor.rr.exceptions.ExceptionCodename;
 import com.scor.rr.exceptions.RRException;
 
 public class DeleteOfLastLayerException extends RRException {
-    public DeleteOfLastLayerException(int id) {
+    public DeleteOfLastLayerException(long id) {
         super(ExceptionCodename.DELETE_LAST_LAYER_IMPOSSIBLE, "Impossible delete of the last Layer with id: " + id );
     }
 }
